Update README.md
Build a music visualizer using a UNO R3 which is a microcontroller board and using arduino to program it. To place the visualizer we made a box in solidworks and laser cut it. With this project we had some issues with the box since they were not cut properly, at then end I gave the idea to cut the pieces that were in the way and it worked. This project was a lot of trial and error since we had to figure out which connections were needed.